Working CC-Match Loops and CT-Match Loops
I have appended the cc-match and ct-match loops to the code. This operation works, but it is only error-free if the files are the same length. It will work if the files are different lengths, but it will print the error "Use of uninitialized value within @Bisseqarray in string eq at Pseudocode.pl line 112" or something similar. Keep this in mind if you test the code.